requirements.txt
# Core
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Telegram
aiogram==3.3.0

# AI/ML
openai==1.6.1
pinecone-client==3.0.0
httpx>=0.25.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Utils
requests==2.31.0
aiohttp==3.9.1
